P4-3091 moving location domain entities into the Domain level package.
