+++ src/main/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/move/Journey.kt
@@ -2,6 +2,7 @@ package uk.gov.justice.digital.hmpps.pecs.jpc.move
 
 import com.beust.klaxon.Json
 import com.beust.klaxon.Klaxon
+import uk.gov.justice.digital.hmpps.pecs.jpc.domain.location.LocationType
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.Event
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.EventDateTime
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.JourneyStateParser
@@ -9,7 +10,6 @@ import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.SupplierParser
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.dateTimeConverter
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.journeyStateConverter
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.supplierConverter
-import uk.gov.justice.digital.hmpps.pecs.jpc.location.LocationType
 import uk.gov.justice.digital.hmpps.pecs.jpc.price.Money
 import uk.gov.justice.digital.hmpps.pecs.jpc.price.Supplier
 import java.time.LocalDateTime
