+++ src/main/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/move/JourneyWithPrice.kt
@@ -1,6 +1,6 @@
 package uk.gov.justice.digital.hmpps.pecs.jpc.move
 
-import uk.gov.justice.digital.hmpps.pecs.jpc.location.LocationType
+import uk.gov.justice.digital.hmpps.pecs.jpc.domain.location.LocationType
 import uk.gov.justice.digital.hmpps.pecs.jpc.price.Money
 
 data class JourneyWithPrice(
