+++ src/main/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/move/Move.kt
@@ -2,6 +2,7 @@ package uk.gov.justice.digital.hmpps.pecs.jpc.move
 
 import com.beust.klaxon.Json
 import com.beust.klaxon.Klaxon
+import uk.gov.justice.digital.hmpps.pecs.jpc.domain.location.LocationType
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.Event
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.EventDate
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.EventDateTime
@@ -13,7 +14,6 @@ import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.dateConverter
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.dateTimeConverter
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.moveStatusConverter
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.supplierConverter
-import uk.gov.justice.digital.hmpps.pecs.jpc.location.LocationType
 import uk.gov.justice.digital.hmpps.pecs.jpc.price.Money
 import uk.gov.justice.digital.hmpps.pecs.jpc.price.Supplier
 import java.time.LocalDate
