+++ src/main/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/price/PriceRepository.kt
@@ -1,7 +1,7 @@
 package uk.gov.justice.digital.hmpps.pecs.jpc.price
 
 import org.springframework.data.jpa.repository.JpaRepository
-import uk.gov.justice.digital.hmpps.pecs.jpc.location.Location
+import uk.gov.justice.digital.hmpps.pecs.jpc.domain.location.Location
 import java.util.UUID
 import java.util.stream.Stream
