+++ src/main/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/service/BasmClientApiService.kt
@@ -11,7 +11,7 @@ import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
 import org.springframework.web.reactive.function.client.WebClient
-import uk.gov.justice.digital.hmpps.pecs.jpc.location.LocationType
+import uk.gov.justice.digital.hmpps.pecs.jpc.domain.location.LocationType
 import java.time.Duration
 import java.time.LocalDate
