+++ src/test/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/auditing/MapLocationMetadataTest.kt
@@ -3,8 +3,8 @@ package uk.gov.justice.digital.hmpps.pecs.jpc.auditing
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.Test
-import uk.gov.justice.digital.hmpps.pecs.jpc.location.Location
-import uk.gov.justice.digital.hmpps.pecs.jpc.location.LocationType
+import uk.gov.justice.digital.hmpps.pecs.jpc.domain.location.Location
+import uk.gov.justice.digital.hmpps.pecs.jpc.domain.location.LocationType
 
 internal class MapLocationMetadataTest {
   @Test
