+++ src/test/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/controller/LocationHistoryDtoTest.kt
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test
 import uk.gov.justice.digital.hmpps.pecs.jpc.auditing.AuditEvent
 import uk.gov.justice.digital.hmpps.pecs.jpc.auditing.AuditEventType
 import uk.gov.justice.digital.hmpps.pecs.jpc.auditing.MapLocationMetadata
-import uk.gov.justice.digital.hmpps.pecs.jpc.location.LocationType
+import uk.gov.justice.digital.hmpps.pecs.jpc.domain.location.LocationType
 import java.time.LocalDateTime
 
 internal class LocationHistoryDtoTest {
