+++ src/test/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/importer/report/TestReportFactory.kt
@@ -1,7 +1,7 @@
 package uk.gov.justice.digital.hmpps.pecs.jpc.importer.report
 
-import uk.gov.justice.digital.hmpps.pecs.jpc.location.Location
-import uk.gov.justice.digital.hmpps.pecs.jpc.location.LocationType
+import uk.gov.justice.digital.hmpps.pecs.jpc.domain.location.Location
+import uk.gov.justice.digital.hmpps.pecs.jpc.domain.location.LocationType
 import uk.gov.justice.digital.hmpps.pecs.jpc.move.Journey
 import uk.gov.justice.digital.hmpps.pecs.jpc.move.JourneyState
 import uk.gov.justice.digital.hmpps.pecs.jpc.move.Move
