+++ src/test/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/integration/pages/MapLocationPage.kt
@@ -3,7 +3,7 @@ package uk.gov.justice.digital.hmpps.pecs.jpc.integration.pages
 import org.fluentlenium.core.annotation.PageUrl
 import org.fluentlenium.core.domain.FluentWebElement
 import org.openqa.selenium.support.FindBy
-import uk.gov.justice.digital.hmpps.pecs.jpc.location.LocationType
+import uk.gov.justice.digital.hmpps.pecs.jpc.domain.location.LocationType
 
 @PageUrl("http://localhost:8080/map-location/{agencyId}")
 class MapLocationPage : ApplicationPage() {
