+++ src/test/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/move/MovePersisterTest.kt
@@ -10,6 +10,7 @@ import org.springframework.context.annotation.Import
 import org.springframework.test.context.ActiveProfiles
 import uk.gov.justice.digital.hmpps.pecs.jpc.TestConfig
 import uk.gov.justice.digital.hmpps.pecs.jpc.config.TimeSource
+import uk.gov.justice.digital.hmpps.pecs.jpc.domain.location.LocationRepository
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.EventType
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.GNICourtLocation
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.journeyEventFactory
@@ -17,7 +18,6 @@ import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.moveEventFactory
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.reportJourneyFactory
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.reportMoveFactory
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.toCourtNomisAgencyId
-import uk.gov.justice.digital.hmpps.pecs.jpc.location.LocationRepository
 import uk.gov.justice.digital.hmpps.pecs.jpc.price.Price
 import uk.gov.justice.digital.hmpps.pecs.jpc.price.PriceRepository
 import uk.gov.justice.digital.hmpps.pecs.jpc.price.Supplier
