+++ src/test/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/move/TestMoveFactory.kt
@@ -1,12 +1,12 @@
 package uk.gov.justice.digital.hmpps.pecs.jpc.move
 
+import uk.gov.justice.digital.hmpps.pecs.jpc.domain.location.LocationType
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.Event
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.EventType
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.Person
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.Profile
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.defaultDate
 import uk.gov.justice.digital.hmpps.pecs.jpc.importer.report.defaultDateTime
-import uk.gov.justice.digital.hmpps.pecs.jpc.location.LocationType
 import uk.gov.justice.digital.hmpps.pecs.jpc.price.Supplier
 import uk.gov.justice.digital.hmpps.pecs.jpc.price.effectiveYearForDate
 import java.time.LocalDate
