+++ src/test/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/service/BasmClientApiServiceTest.kt
@@ -12,7 +12,7 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
 import org.springframework.web.reactive.function.client.WebClient
-import uk.gov.justice.digital.hmpps.pecs.jpc.location.LocationType
+import uk.gov.justice.digital.hmpps.pecs.jpc.domain.location.LocationType
 import java.time.Duration
 import java.time.LocalDate
 import java.util.concurrent.TimeUnit
